Refactor kernel
