+++ burn-wgpu/src/compute/mod.rs
@@ -1,7 +1,9 @@
 mod base;
+mod kernel;
 mod server;
 mod storage;
 
 pub use base::*;
+pub use kernel::*;
 pub use server::*;
 pub use storage::*;
